docs: wire POS content and enforce snippet checks
