+++ docs/.markdown-link-check.json
@@ -1,0 +1,6 @@
+{
+  "ignorePatterns": [
+    { "pattern": "^mailto:" },
+    { "pattern": "^tel:" }
+  ]
+}
